Updates question mock data and their types
